+++ src/mods/puzzle/app/canisters/plugin.tsx
@@ -12,6 +12,7 @@ const CanistersComponent: FamilyPlugin<CanistersPuzzleData>["Component"] = ({ pu
     difficulty={ctx.difficulty}
     role={ctx.role}
     theme={ctx.theme}
+    room={ctx.address}
     onSolved={onSolved}
     onCancel={onCancel}
   />